# api/openapi.yaml
openapi: 3.0.3
info:
  title: Office Tracker API
  description: |
    REST API for tracking office attendance using RFID cards.

    ## Authentication
    If API key authentication is enabled (via `API_KEY` environment variable),
    all endpoints except `/health` require authentication using one of these methods:
    - Header: `X-API-Key: your-api-key`
    - Header: `Authorization: Bearer your-api-key`

    ## CORS
    CORS can be enabled by setting the `CORS_ORIGINS` environment variable.
    - Set to `*` to allow all origins
    - Set to comma-separated list for specific origins (e.g., `http://localhost:3000,https://example.com`)
  version: 1.0.0
  contact:
    name: Office Tracker Team
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/sessions/checkin:
    post:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Check in a user
      description: Create a new active session for a user
      tags:
        - Sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSessionActionRequest"
      responses:
        "200":
          description: User checked in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSessionActionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/sessions/checkout:
    post:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Check out a user
      description: Close the active session for a user
      tags:
        - Sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSessionActionRequest"
      responses:
        "200":
          description: User checked out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSessionActionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/users:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: List and search users
      description: Retrieve a list of users with optional search, sorting, and pagination
      tags:
        - Users
      parameters:
        - name: search
          in: query
          required: false
          description: Search users by name (partial match)
          schema:
            type: string
            example: John
        - name: limit
          in: query
          required: false
          description: Maximum number of results (default 50)
          schema:
            type: integer
            minimum: 1
            example: 50
        - name: offset
          in: query
          required: false
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            example: 0
        - name: order
          in: query
          required: false
          description: Sort order (default 'asc')
          schema:
            type: string
            enum: [asc, desc]
            example: asc
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (default 'name')
          schema:
            type: string
            enum: [name, created_at]
            example: name
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Create a new user
      description: Register a new user in the system
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Delete users (bulk)
      description: |
        Delete users matching filters.
        Requires at least one filter (currently `search`) to prevent deleting all users.
      tags:
        - Users
      parameters:
        - name: search
          in: query
          required: false
          description: Delete users with matching names (partial match)
          schema:
            type: string
            example: John
        - name: order
          in: query
          required: false
          description: Sort order for filtered delete candidates
          schema:
            type: string
            enum: [asc, desc]
            example: asc
      responses:
        "200":
          description: Users deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                    format: int64
                    example: 3
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/users/export:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Export users as CSV
      description: Export all users to a CSV file
      tags:
        - Users
      parameters:
        - name: search
          in: query
          required: false
          description: Search users by name (partial match)
          schema:
            type: string
            example: John
        - name: order
          in: query
          required: false
          description: Sort order (default 'asc')
          schema:
            type: string
            enum: [asc, desc]
            example: asc
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (default 'name')
          schema:
            type: string
            enum: [name, created_at]
            example: name
        - name: limit
          in: query
          required: false
          description: Maximum number of results (default 50)
          schema:
            type: integer
            minimum: 1
            example: 50
        - name: offset
          in: query
          required: false
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            example: 0
      responses:
        "200":
          description: CSV file with all users
          content:
            text/csv:
              schema:
                type: string
                example: |
                  Name,RFID_UID,DiscordID
                  John Doe,ABC123456,123456789012345678
                  Jane Smith,XYZ789012,987654321098765432
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/users/import:
    post:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Import users from CSV
      description: Import users from a CSV file (Name,RFID_UID,DiscordID)
      tags:
        - Users
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file to upload
              required:
                - file
      responses:
        "200":
          description: Import completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                    description: Number of users successfully imported
                    example: 10
                  failed:
                    type: integer
                    description: Number of users that failed to import
                    example: 2
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/presence:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Get active presence
      description: Retrieve all currently active users in the office
      tags:
        - Sessions
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PresenceResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/sessions:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: List sessions with filters
      description: Retrieve sessions with optional filtering by date range, user name, sorting, and pagination
      tags:
        - Sessions
      parameters:
        - name: from
          in: query
          required: false
          description: Filter sessions from this datetime (RFC3339 format)
          schema:
            type: string
            format: date-time
            example: "2026-02-01T00:00:00Z"
        - name: to
          in: query
          required: false
          description: Filter sessions until this datetime (RFC3339 format)
          schema:
            type: string
            format: date-time
            example: "2026-02-28T23:59:59Z"
        - name: name
          in: query
          required: false
          description: Filter sessions by user name (partial match)
          schema:
            type: string
            example: John
        - name: user_id
          in: query
          required: false
          description: Filter sessions by user ID
          schema:
            type: integer
            format: int64
            example: 1
        - name: discord_id
          in: query
          required: false
          description: Filter sessions by Discord ID
          schema:
            type: string
            example: "123456789012345678"
        - name: status
          in: query
          required: false
          description: Session status filter
          schema:
            type: string
            enum: [all, active, completed]
            example: completed
        - name: active_only
          in: query
          required: false
          description: Legacy status alias (`true` = active, `false` = completed)
          schema:
            type: boolean
            example: false
        - name: limit
          in: query
          required: false
          description: Maximum number of results (default 100)
          schema:
            type: integer
            minimum: 1
            example: 100
        - name: offset
          in: query
          required: false
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            example: 0
        - name: order
          in: query
          required: false
          description: Sort order (default 'desc')
          schema:
            type: string
            enum: [asc, desc]
            example: desc
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (default 'check_in')
          schema:
            type: string
            enum: [check_in, check_out, user_name]
            example: check_in
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Delete sessions (bulk)
      description: |
        Delete sessions matching filters.
        At least one filter must be provided.
        When `limit` or `offset` are supplied, only that ordered slice is deleted.
      tags:
        - Sessions
      parameters:
        - name: from
          in: query
          required: false
          description: Delete sessions from this datetime (RFC3339)
          schema:
            type: string
            format: date-time
            example: "2026-02-01T00:00:00Z"
        - name: to
          in: query
          required: false
          description: Delete sessions until this datetime (RFC3339)
          schema:
            type: string
            format: date-time
            example: "2026-02-28T23:59:59Z"
        - name: name
          in: query
          required: false
          description: Delete sessions by user name (partial match)
          schema:
            type: string
            example: John
        - name: user_id
          in: query
          required: false
          description: Delete sessions for a specific user ID
          schema:
            type: integer
            format: int64
            example: 1
        - name: discord_id
          in: query
          required: false
          description: Delete sessions for a specific Discord ID
          schema:
            type: string
            example: "123456789012345678"
        - name: status
          in: query
          required: false
          description: Delete only sessions with the selected status
          schema:
            type: string
            enum: [all, active, completed]
            example: completed
        - name: active_only
          in: query
          required: false
          description: Legacy status alias (`true` = active, `false` = completed)
          schema:
            type: boolean
            example: false
        - name: limit
          in: query
          required: false
          description: Maximum number of matching rows to delete from the ordered result set
          schema:
            type: integer
            minimum: 1
            example: 100
        - name: offset
          in: query
          required: false
          description: Number of matching rows to skip before deleting
          schema:
            type: integer
            minimum: 0
            example: 0
        - name: order
          in: query
          required: false
          description: Sort order applied before pagination (default `desc`)
          schema:
            type: string
            enum: [asc, desc]
            example: desc
        - name: sort_by
          in: query
          required: false
          description: Field used for ordered bulk delete pagination
          schema:
            type: string
            enum: [check_in, check_out, user_name]
            example: check_in
      responses:
        "200":
          description: Sessions deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                    format: int64
                    example: 15
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/sessions/count:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Count sessions with filters
      description: Retrieve the total number of sessions matching optional filters
      tags:
        - Sessions
      parameters:
        - name: from
          in: query
          required: false
          description: Filter sessions from this datetime (RFC3339 format)
          schema:
            type: string
            format: date-time
            example: "2026-02-01T00:00:00Z"
        - name: to
          in: query
          required: false
          description: Filter sessions until this datetime (RFC3339 format)
          schema:
            type: string
            format: date-time
            example: "2026-02-28T23:59:59Z"
        - name: name
          in: query
          required: false
          description: Filter sessions by user name (partial match)
          schema:
            type: string
            example: John
        - name: user_id
          in: query
          required: false
          description: Filter sessions by user ID
          schema:
            type: integer
            format: int64
            example: 1
        - name: discord_id
          in: query
          required: false
          description: Filter sessions by Discord ID
          schema:
            type: string
            example: "123456789012345678"
        - name: status
          in: query
          required: false
          description: Session status filter
          schema:
            type: string
            enum: [all, active, completed]
            example: active
        - name: active_only
          in: query
          required: false
          description: Legacy status alias (`true` = active, `false` = completed)
          schema:
            type: boolean
            example: false
      responses:
        "200":
          description: Session count retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    format: int64
                    example: 248
                required:
                  - total
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/sessions/export:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Export sessions as CSV
      description: Export sessions to a CSV file with optional date filtering
      tags:
        - Sessions
      parameters:
        - name: from
          in: query
          required: false
          description: Filter sessions from this datetime (RFC3339 format)
          schema:
            type: string
            format: date-time
            example: "2026-02-01T00:00:00Z"
        - name: to
          in: query
          required: false
          description: Filter sessions until this datetime (RFC3339 format)
          schema:
            type: string
            format: date-time
            example: "2026-02-28T23:59:59Z"
        - name: name
          in: query
          required: false
          description: Filter sessions by user name (partial match)
          schema:
            type: string
            example: John
        - name: user_id
          in: query
          required: false
          description: Filter sessions by user ID
          schema:
            type: integer
            format: int64
            example: 1
        - name: discord_id
          in: query
          required: false
          description: Filter sessions by Discord ID
          schema:
            type: string
            example: "123456789012345678"
        - name: status
          in: query
          required: false
          description: Session status filter
          schema:
            type: string
            enum: [all, active, completed]
            example: active
        - name: active_only
          in: query
          required: false
          description: Legacy status alias (`true` = active, `false` = completed)
          schema:
            type: boolean
            example: false
        - name: limit
          in: query
          required: false
          description: Maximum number of results (default 100)
          schema:
            type: integer
            minimum: 1
            example: 100
        - name: offset
          in: query
          required: false
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            example: 0
        - name: order
          in: query
          required: false
          description: Sort order (default 'desc')
          schema:
            type: string
            enum: [asc, desc]
            example: desc
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (default 'check_in')
          schema:
            type: string
            enum: [check_in, check_out, user_name]
            example: check_in
      responses:
        "200":
          description: CSV file with sessions
          content:
            text/csv:
              schema:
                type: string
                example: |
                  UserName,CheckIn,CheckOut,Duration(minutes)
                  John Doe,2026-02-15T09:00:00Z,2026-02-15T17:00:00Z,480.00
                  Jane Smith,2026-02-15T10:00:00Z,,
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/sessions/open:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Get all open sessions
      description: |
        Retrieve all currently open (not checked out) sessions with user information.
        Returns SessionWithUser objects including user_name from JOIN query.
      tags:
        - Sessions
      responses:
        "200":
          description: Successful response with array of SessionWithUser objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
              example:
                - id: 1
                  user_id: 1
                  user_name: John Doe
                  check_in: "2026-02-15T09:00:00Z"
                  check_out: null
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/sessions/checkout-all:
    post:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Check out all active sessions
      description: Sign out all currently checked-in members at once (bulk checkout)
      tags:
        - Sessions
      responses:
        "200":
          description: Bulk checkout completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully checked out 5 session(s)"
                  count:
                    type: integer
                    example: 5
                  sessions:
                    type: array
                    items:
                      type: string
                    example:
                      - "John Doe"
                      - "Jane Smith"
                      - "Bob Johnson"
                      - "Alice Brown"
                      - "Charlie Davis"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/environment:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Get latest environmental reading
      description: Retrieve the latest in-memory temperature reading together with freshness metadata.
      tags:
        - Environment
      responses:
        "200":
          description: Latest environmental reading metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
    post:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Update latest environmental reading
      description: Store the latest temperature reading in memory. Intended for ESP32 sensor updates.
      tags:
        - Environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvironmentUpsertRequest"
      responses:
        "200":
          description: Environmental reading stored successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/statistics/leaderboard:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Get leaderboard
      description: Retrieve leaderboard statistics for a metric over a date range
      tags:
        - Statistics
      parameters:
        - name: rank_by
          in: query
          required: false
          description: Preferred metric to rank the leaderboard by
          schema:
            type: string
            enum: [hours, visits]
            example: hours
        - name: limit
          in: query
          required: false
          description: Maximum results (default 10)
          schema:
            type: integer
            minimum: 1
            example: 10
        - name: from
          in: query
          required: false
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            example: 2026-02-01
        - name: to
          in: query
          required: false
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            example: 2026-02-08
        - name: include_auto_checkout
          in: query
          required: false
          description: "Include non-RFID checkouts (including 04:00 auto-checkout) (default: false)"
          schema:
            type: boolean
            example: false
      responses:
        "200":
          description: Leaderboard results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/reports/weekly:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Get latest weekly report
      description: Retrieve the latest generated weekly activity report with trends and leaderboards
      tags:
        - Reports
      responses:
        "200":
          description: Weekly report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReport"
        "404":
          description: No report available yet
          content:
            text/plain:
              schema:
                type: string
                example: no report available
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (optional, configured via API_KEY env var)
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication (alternative to X-API-Key)
  schemas:
...
    UserSessionActionRequest:
      type: object
      description: Request body for checking in/out a user
      properties:
        user_id:
          type: integer
          format: int64
          example: 1
      required:
        - user_id
    UserSessionActionResponse:
      type: object
      properties:
        status:
          type: string
          example: checked-in
      required:
        - status
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        text/plain:
          schema:
            type: string
            example: invalid request
...
    InternalServerError:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
            example: database error